Add module list commands to arkouda cray-cs and cray-xc jobs
In debugging issues with the arkouda jobs as a result of our Python 3 changes,
I found myself wanting to know more about the environment in which these jobs
were run.  This change adds two `module list` commands to the jobs - one prior
to any modifications and one just after we expect the module list to be stable.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>