Stop loading a gcc "gen" compiler for non gnu builds
We used to load a gcc gen compiler so that we'd have a newer gcc to
build the compiler with even when doing intel/cce runtime builds.
However, the Lmod module system does not allow intel and gcc to be
loaded at the same time so stop doing this. This was motivated by gcc
7.3.0 not being available on our target machine anymore, but even if we
used a newer version the `module load PrgEnv-intel` will just unload
gcc.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>