Resolves https://github.com/Cray/chapel-private/issues/1392 Fixes memory leak in `test/reductions/reduceLoopOfPromoted.chpl`
This applies the solution for avoiding memory leaks, discovered in #16518
for the then-added chpl_elemTypeForReducingIterables(), to iteratorIndexType().
Moves the explanatory comments there as well.

While there, splits a couple of long lines, which I was supposed
to do in #16518.

Does not affect other known memory leaks.

Future work: apply the same solution to iteratorIndex() itself.
For that, it needs to be passed the iterator itself
rather than an iterator class (as it is done now).

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>