Change to creating a chpldeps "Python app"
Started to do this as a zipapp but sphinx doesn't like
to exist within a zipfile (we get errors when it tries to
open a theme). So now we just install the dependencies
and extend the python path to include them.

For convenience, the 'chpldeps' Python app will run
included dependencies, e.g.
  chpldeps sphinx-build arg...

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>