Move the inheritance stuff into a recursive function
This solves the issue with not going up into more than just the immediate parent
type.

I maybe could condense it more, to reuse all the code.  However, doing so would
mean repetitive checks when we know we have AggregateTypes already.  I dunno if
it makes that much of a difference, though.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>