Futurize canSeeSecondaryMethod2.chpl
This test tracks being able to find methods on a type when the type is listed in
an only list.  Because we no longer ignore the privacy and limitations of use
statements, this currently has stopped working.  However, we intend to allow
specifying the type to impact the methods defined on that type visible from that
scope, so this should not be a permanent change.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>