Don't default to CHPL_GMP=system with PrgEnv-cray on Cray EX systems.
The Chapel module has been defaulting to CHPL_GMP=system with the
PrgEnv-cray target compiler, but there isn't actually a system gmp on EX
systems, so don't do that there.

(Based on a conversation with another developer, this was originally
done because the CCE compiler couldn't build gmp, possibly because of
the inline assembly.  So, we had to use the system one.  But that's no
longer true, so when it comes time to cherry-pick all the module-related
changes I'm working on back to master I'll just undo this entirely.  For
now, though, I want to limit myself to only those changes which affect
use of the EX module.)

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>