Fix bug with not finding symbol imported in generic function
It turns out that you can't compare astr("") to itself - doing so caused
ImportStmts to fail to copy the unqualified and renaming lists, which meant
generic instantiations couldn't utilize symbols brought in for unqualified
access via an import statement.  Replace comparing to astr("") with comparing
the first letter to '\0'.

While here, I noticed that moving the new name when we notice the import is of
a symbol in a module was not using astr to replace it.  Which isn't that
important, since as I said earlier you can't compare two astrs of "".  But it's
good to have consistency so we don't accidentally start failing to use astr on
something other than "".

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>