Convert chpl_nodeID to an `extern proc`
Declaring `chpl_nodeID` as a module-scope extern variable pointed to
an issue in which we weren't introducing a shadow variable for it as
we do for other module-scope variables.  However, it also shows a
weakness of using a module-scope extern variable for an external value
that is different on each locale: it suggests to Chapel that there's a
single extern variable when in fact there's a per-locale variable with
a distinct value on each locale.  This suggests that an extern var
isn't the best way to represent such values in Chapel.

In this PR, I change chpl_nodeID to an extern (paren-less) proc so
that the Chapel code which refers to it can remain unchanged; and I
assign it a new name in the C code so that we can implement it as a
static inline function without modifying other C code that refers to
the C-level variable.  The result is a minimal change to our code that
avoids the mismatch between describing such a value as a global
variable in Chapel that should result in a nearly identical
execution-time implementation.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>