Fix memory leak for global array type expressions
This commit fixes the following so it does not leak:

  type arrOfArr = [1..2][100..200] int;
  writeln(arrOfArr:string);

This is visible in the existing test
  types/typedefs/bradc/arrayTypedef.chpl

The problem here is that evaluateAutoDestroy in normalize.cpp
hoists some temporaries out of forall-expr functions representing
types into module scope. Meanwhile, the change to callDestructors in
PR #16430 caused it to no longer deinitialize such globals (because
it never found the initialization because it is within another function).
This commit adjusts collectGlobals in callDestructors to delve into
functions that might return array types.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>