Remove a local block from domain's array tracking code
With 16740 we're now accessing remote data in what was a local block so
remove the local block. The alternative would be change to an external
list that localizes references, but that introduces allocations and
other overheads in the common case, whereas this introduces comm in a
relatively uncommon case of destroying an array on a different node than
the domain or other arrays declared over it.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>