Change CHPL_LLVM=llvm to CHPL_LLVM=bundled everywhere.
This follows up https://github.com/chapel-lang/chapel/pull/16540.  It
converts all CHPL_LLVM=llvm settings to CHPL_LLVM=bundled.  We had
thought we didn't need to do that because the original PR caused
'llvm' to be treated as 'bundled' implicitly, but it also caused a
warning to be printed and that warning led to some unexpected
problems.  So here, just convert all the =llvm settings to =bundled.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>