Hard-code aliases into third-party-pkgs along with their env-var names
Since multiple packages can use the 'bundled' name for their setting type,
we need to pay attention to both the env-var name and the value. Hard-code
a list of varname-value pairs to map to directories instead of pulling them
out of third-party/*/Makefile.include. Look them up based on varname-value.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>