Replace chplenv variables in ReadFileWithComments
`ReadFileWithComments` and `runSkipIf` are responsible for reading most
testing infrastructure files in sub_test. Both can either just read in a
non-executable file or run an executable file and read in the output.
`runSkipIf` will inject the chplenv for executable and non-executable
files, but `ReadFileWithComments` was only doing so for executable
files.

For non-executable files `ReadFileWithComments` was only injecting
global env vars. Update it to inject chplenv vars as well.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>