Convert the protobuf EXECENV into a script to hopefully allow it to see the var
Elliot suspected that we'd see a similar error to the skipifs with the EXECENV,
we just hadn't gotten there because the skipifs failed first.  Unfortunately,
we don't bring the chplenv vars into EXECENVs.  This commit adjusts the
EXECENV to instead follow the lead of the EXECENV in test/mason/mason-doc, to
bring in the right installation location to look.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>