Clean up expandvars_chpl implementation in sub_test
Use string.Template.safe_substitute instead of expandvars to avoid
having to manipulate os.environ. Note that we use safe_substitute
instead of substitute so that if a var isn't set we don't throw an
exception. This is needed for tests like test/mason/masonNewBadName.chpl
and matches the behavior of expandvars.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>