Fix an assertion error that comes up when programs have syntax errors
This fixes an assertion error pointed out in an SO issue:

https://stackoverflow.com/questions/65137722/why-did-i-get-this-behaviour/65139736#65139736

that occurs when the body of a 'cobegin' contains a syntax error.
Here, I'm working around it by replacing the assertion with a test and
returning early in such cases, believing the compiler will exit before
the lack of a built cobegin statement catches up with us.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>