Make user-defined '==' and '!=' overloads on classes illegal
Since we've made '=' illegal on classes, we should really make '=='
and '!=' illegal as well since they are symmetric with '=' and as such
are the fundamental ways of checking the class pointer/reference value
that is being stored (e.g., to compare against `nil`).  We've
discussed users who want to compare classes defining a .equals()
method for such purposes.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>