Call setSubprogram on debug info for function when it is created
When doing codegen for a function with --lldb and -g, we were creating a
llvm::Subprogram unit for debug scope information, but never actually
assigning it to the function. Call setSubprogram with the debug scope
subprogram in order to pair it with the function.

Fixes execflags/bradc/gdbdash/declint and execflags/bradc/gdbdash/gdbSetConfig
with --llvm and --fast.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>