Add a use of the IO module to the Random modules to enable calls to <~>
The writeThis methods in the Random module were getting away with calls to <~>,
probably due to uses of the IO module at the points where it is called.
Unfortunately, a user encountered a situation where such a use was not
available, causing the writeThis to fail to resolve. This adds a use of the IO
module at the scopes with the writeThis functions, to avoid this issue in the
future.

Passed a full std paratest with futures

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>