Use fork to isolate GPU code generation
We are seeing some sort of problem where the LLVM verifier reports
problems for the code CPU generated after the GPU code is generated.
There might be some global state that is not cleared properly
after the GPU code generation. As a workaround, run fork()
and do the GPU code generation in a separate process.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>