Prefer PMI2 to PMIx for slurm integration with comm=ofi out-of-band.
Based on limited experience on the CS platform we've only supported PMIx
for integrating the Slurm and the comm=ofi out-of-band support.  But it
turns out that PMI2 can also be found on CS sometimes, and that seems a
better solution if its client-side library is supplied by Slurm itself.
Here, when that is the case, use it instead.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>