Quiet some tests under gasnet ASan
Skip --cache-remote variant of optimizations/sungeun/optimized-on/loop when
ASan is on (cache-remote is not currently compatible with ASan)

Skip interop/C/multilocale/globals/use_reliesOnGlobal unless memory sanitizer
is on. This test relies on an uninitialized global having a defined value,
which is undefined behavior. The value is some sentinel under ASan, and happens
to be 0 on most systems we run on, but the behavior is truly undefined, so
don't run it unless memory sanitizer is on. We don't actually have memory
sanitizer testing yet, but when we do this test will run and should get a real
error message from MSan.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>