appendExpr(): Improve handling of operators
Handle prefix-~ and postfix-!

If the two-actual expr's name starts with a letter (or _), don't print
it as an operator.  Treat it the same as an expr with other than two
actuals.

Also tidy the logic so that two-arg non-operators use the same else
block as not-two-arg cases.

Add three tests locking in the changed behavior.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>