Restrict 'sort()' to only support 1D rectangular arrays
In issue #16888 a user pointed out that our sort() routine currently
accepts sparse arrays, but then doesn't really do anything useful
with them (it was not intended to support them).  Previously, an
error message was generated for multidimensional arrays.  This
extends that error message to also support non-rectangular arrays
and updates the message slightly to reflect that.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>