Ease associating CQ errors with the operations that caused them.
We weren't reporting the operation context value when a CQ reported an
error entry.  And when we didn't need to track the completion of an
operation we were just giving a null pointer as the context value.  Both
of those made it harder to figure out which operation had failed, when
such an error entry occurred.  Here, start reporting the context when
errors occur, and use the source line number as a context value when we
don't need to track the operation's completion explicitly.

While here, also simplify creating transaction tracking context values,
and fix a bug in how we reported the provider string value for CQ error
entries.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>