Always enable the unused-operand1 workaround.
We've had a RxD-specific workaround to supply a pointer for the operand1
argument to a fetching AMO even when the AMO operation was a read.  In
the absence of that workaround the RxD provider would return -FI_EINVAL
from fi_fetch_atomic(op_read) when operand1 was NULL, even though the
man page says that's allowed.  Now we've seen another provider behave
the same way, so just enable the workaround in all cases.  It doesn't do
any harm.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>