Increase the problem size for the multi-locale scan performance test
The old problem size was set when distributed scan was serial and slow,
but it was significantly optimized a while back in 12481. The current
problem size finishes in ~0.01 seconds, which isn't that telling.
Increase the problem size to 1/4 of memory, which is what we use for the
single locale problem size.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>