Make verbose flag tests pass in more HPE Cray EX environments.
Adjust so that testVFlag and testVerbose flag will pass with or without
an explicit Slurm partition specified when that launcher is used.  Add
HPE Cray EX systems to the list of those where the compute node names
should be ignored because they're not predictable.  We ran into both of
these while testing on a new-to-us in-house EX system.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>