Rename test/types/locale/hostname.chpl to test-hostname.chpl
The test was named hostname but also used the `hostname` command to generate
its .good file which was a little confusing. If run with ./ in $PATH, the
test's .prediff called the test itself instead of the system command. In
a multilocale setting with ./ in $PATH, it failed because the .prediff didn't
use the --numLocales option when running what it thought was the system
hostname.

Update the prediff and .gitignore file to use the new test name.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>