Add support for naming the type in the only list during visible functions
Fleshes out the typeWasNamed method on use statements for both only and except
lists, and calls the method when finding all the visible functions.

This resolves the only list tests.  There is still something off for the except
list tests, will investigate (but am committing the other changes for now)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>