Don't add pmi to CHPL_LD_FLAGS in ofi testing if pmi is a system library
On our new CS machine, slurm/pmi are system libraries and are found by
default by the linker. Previously, we were adding `-L /usr/lib64` to
`CHPL_LD_FLAGS`, which would result in the system libfabric being
brought in and its not compatible with the bundled version we're trying
to use since 16968.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>