Make python2 available for arkouda release testing on cray-cs
Arkouda release testing uses chapel 1.23, which requires python2 and for
python to be a python2. On CS systems we used to do this by loading a
python 2 module, but I thought that was no longer needed and removed it
in 16967. However, it is still needed for arkouda release testing. Use
that module for arkouda release testing on cray-cs-hdr, but use a
different trick (sourcing a PATH that has a python symlinked to
/usr/bin/python2) to bring in python2 for arkouda release testing on the
new cray-cs system that does not have the module available.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>