Try to ensure arkouda dependency versions match our test dependencies
We pin to PyYaml 5.3.1 in test-requirements, but an Arkouda dependency
was pulling in 5.4.0 because we throw `--upgrade --no-cache-dir` since
15654. There is some issue in PyYaml 5.4 that is breaking yaml imports
at the moment. To fix this and hopefully avoid similar issues in the
future list our test-requirements as a constraint for versioning. This
should make sure that Arkouda dependencies pull in the same versions we
want, assuming they don't hard code versions themselves.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>