Improve how CHPL_CACHE_REMOTE is set
This is a module param that's set by the compiler. Change it to be more
like other compiler set params such as `boundsChecking`. Previously the
value was only changed when `--[no-]cache-remote` was thrown on the
command line, but as we look enable the cache by default we want to
always set the value.

See https://github.com/Cray/chapel-private/issues/1602 for context

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>