CG: Implicit inference of implements-statements
Infer an `implements` statement, if possible, when there is no explicit
`implements` statement to satisfy an interface constraint upon a call
to a CG function.

The implementation strategy is to have the compiler create such a
statement under those circumstances and attempt to resolve it.
As a caching strategy, leave it in the AST in either case so that
later inference attempts can find and reuse it. Mark it with prim_error
to indicate unsuccessful resolution attempt.

This implements the semantics presented in #16994 as of this writing.
Given its experimental nature, it is controlled by a developer flag,
which is enabled by default: --infer-implements-decls

The case of nested implementations is currently not working correctly,
as captured in `infer-implements-stmt.future`

While there: remove an unnecessary restriction on the actual arguments
of `implements` statements, clean up the license text for the two files
added in #16942, add some bug futures.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>