Disable remote cache if address sanitizer is enabled
The cache reads whole cache lines at a time, even if only part of the
page is allocated. This is an intentional design choice, but it is
technically reading unallocated memory, which is one of the things ASan
catches. For now just disable the cache if asan is enabled.

See https://github.com/Cray/chapel-private/issues/933

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>