Simplify "obvious mismatch" logic and improve camelCase
This routine was overly complicated due to assumptions on my part that
it wouldn't be this simple and expectations that I'd make it eliminate
more complicated mismatches.  Turned out it was simple and that I
haven't had sufficient time or motivation to look at the more
complicated cases, so let's make it as simple as it currently is.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>