Stop defaulting to CHPL_GMP=system with PrgEnv-cray.
The Chapel module has been defaulting to CHPL_GMP=system with the
PrgEnv-cray target compiler.  This was originally done because the CCE
compiler couldn't build gmp, probably because of the inline assembly.
So, we had to use the system one.  But on EX systems there is no system
gmp (see 1874ae8) and in any case the modern CCE compiler can build gmp,
so stop defaulting to a system one entirely.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>