Fix memory leak in ReplicatedDist, add a test, adjust a test
Fix a memory leak in ReplicatedDist that was caused by replicated
arrays default-initializing every replicand, even when
`initElts=false`.

Previously, a replicated array would ignore the value of `initElts`.

Now, when a replicated array is initialized with `initElts=false`,
the replicand on the same locale is also initialized with
`initElts=false`, but all other replicands are default-initialized.

This suggests that there is no way to `initElts=false` an entire
replicated array (for N locales, N-1 replicands will always be
default-initialized).

Changing the behavior of `initElts` for replicated was required
because `chpl__coerceCopy()` relies on `initElts=false` when
copy-initializing an array.

This strange behavior for `initElts` is due to how ReplicatedDist
works. It makes no consistency guarantees for replicands on different
locales, starting at initialization time.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>