Retire const-checking futures fixed on Jan 25
Here, I'm retiring some const-checking futures that I filed, and
others in the same directory, and reverting some changes to Rafael
Asenjo's tests which were added at some point as a workaround with
these futures.  (One of the futures also claimed that some changes to
our blackscholes benchmarks could also be reverted, but this turned
out not to be true due to an assignment between records later in the
code.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>