Fix a bug in cleanup logic.
When we revert aggregation because of a failing lifetime analysis, we
need to remove the aggregation code we generate. Before this commit, we
were doing that cleanup with the assumption that an aggregator can be
used in only one location in the AST. However, in the case where the
aggregation is added to a fast/slow follower pair, we end up using the
same aggregator in both follower bodies.

Signed-off-by: Engin Kayraklioglu <e-kayrakli@users.noreply.github.com>