Revert optional changes in modules
This commit reverts the uses of if-var in the modules.
The reason is to avoid the (however unlikely) possibility
of performance impact from #17047. We are expecting a big merge
in the next couple of days and would like to observe its impacts
in isolation.

I intend to revert this commit, i.e. add back those uses of if-var,
on a slow(er) day.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>