Avoid using catfiles for lulesh testing
Add an option to lulesh to print coords to stdout instead of a file and
use that for correctness testing. Some launchers and systems do not
guarantee that files written by a job will be visible on disk by the
time the job completes. This was causing issues for lulesh where it
writes output to `coords.dat` and we then try to `cat` the file to our
output, but the file wasn't visible on disk when we tried to `cat`.

We have seen this in the past on systems using pure aprun (no qsub) and
recently on systems using lsf/bsub.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>