Quiet a Tcl module collection-related message.
There is a rare (we're told) race that causes module collection creation
to fail because it's already been done, in multi-node jobs on systems
using Tcl v4 modules where the user home dir is NFS-mounted. We've seen
a failure caused by this message on our CS system, so here, filter it
out.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>