Don't steal RHS for arrays with different element types
This change addresses a failure with
test/studies/shootout/nbody/sidelnik/nbody_orig_1.chpl.

After PR #16802, array literals are created with `chpl__buildArrayExpr`
using the `in` intent. The result is that the code

  var bodies: [0..#NBODIES] borrowed Planet = [ new Planet(...), ... ]

creates a RHS temporary array of owned - let's call it `tmpArray`.
`tmpArray` is destroyed at the end of the block (and the temporaries for
the individual `new` call results are moot at this point since they are
`nil` from having been transferred out of).

This was working as of PR #16802 because the array `bodies` array was
created by copy-initializing from `tmpArray`. However, without this
change, the `bodies` array is created by move-initializing from
`tmpArray`, which leads to use-after-free.

The fix is to not use move-initialize in insertCasts unless the
compile-time types match.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>