Update and resolve this future now that the feature it tracked was added
This future was requesting the ability to define an assignment operator as a
method on a type.  Now that operators can be defined as methods, it can be
resolved.  However, it needed some updates as a result of other decisions that
have occurred over time.  Namely:
- Classes cannot have assignment operators
- The method form of an assignment operator is not expected to infer the lhs
argument and adjust - operator methods are considered type methods rather than
methods on a particular instance, so it is not reasonable to insert an implicit
this argument and use it.

With these changes, the test now passes and the .future and .bad file can be
removed

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>