Make the path to the MLI REAL_BINARY_NAME explicit
Convert the multi-locale interop binary from `filename` to `./filename`.
Previously, we were just specifying the basename of the file and not
giving an execution location. This used to work when gasnet used `execv`
to launch since `execv` will find executables just based on the name,
but in changing to `execvp` in the 2020.10.0 release this no longer
works for us. We should have always been specifying the execution
location, we just lucked out previously that gasnet was using `execv`.
Longer term we should probably find the absolute path to the filename
like we do for non-mli configurations but in the meantime this gets
testing working again.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>